requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
